+++ jsonschema-examples/src/test/java/com/github/victools/jsonschema/examples/ExampleTest.java
@@ -36,6 +36,7 @@ public class ExampleTest {
             ExternalRefAnnotationExample.class,
             ExternalRefPackageExample.class,
             IfThenElseExample.class,
+            InheritanceRefExample.class,
             JacksonDescriptionAsTitleExample.class,
             JacksonSubtypeDefinitionExample.class,
             StrictTypeInfoExample.class,