+++ jsonschema-module-jackson/src/test/java/com/github/victools/jsonschema/module/jackson/AbstractTypeAwareTest.java
@@ -22,6 +22,7 @@ import com.fasterxml.classmate.members.ResolvedField;
 import com.fasterxml.classmate.members.ResolvedMethod;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.github.victools.jsonschema.generator.FieldScope;
+import com.github.victools.jsonschema.generator.MemberScope;
 import com.github.victools.jsonschema.generator.MethodScope;
 import com.github.victools.jsonschema.generator.SchemaGenerationContext;
 import com.github.victools.jsonschema.generator.SchemaGeneratorConfig;
@@ -40,7 +41,7 @@ public class AbstractTypeAwareTest {
 
     private final Class<?> testClass;
     private SchemaGenerationContext context;
-    private ResolvedTypeWithMembers testClassMembers;
+    private MemberScope.DeclarationDetails declarationDetails;
 
     protected AbstractTypeAwareTest(Class<?> testClass) {
         this.testClass = testClass;
@@ -54,7 +55,7 @@ public class AbstractTypeAwareTest {
     protected void prepareContextForVersion(SchemaVersion schemaVersion) {
         TypeContext typeContext = TypeContextFactory.createDefaultTypeContext();
         ResolvedType resolvedTestClass = typeContext.resolve(this.testClass);
-        this.testClassMembers = typeContext.resolveWithMembers(resolvedTestClass);
+        this.declarationDetails = new MemberScope.DeclarationDetails(resolvedTestClass, typeContext.resolveWithMembers(resolvedTestClass));
         this.context = Mockito.mock(SchemaGenerationContext.class, Mockito.RETURNS_DEEP_STUBS);
         Mockito.when(this.context.getTypeContext()).thenReturn(typeContext);
 
@@ -94,26 +95,26 @@ public class AbstractTypeAwareTest {
     }
 
     protected FieldScope getTestClassField(String fieldName) {
-        ResolvedField resolvedField = Stream.of(this.testClassMembers.getMemberFields())
+        ResolvedField resolvedField = Stream.of(this.declarationDetails.getDeclaringTypeMembers().getMemberFields())
                 .filter(field -> fieldName.equals(field.getName()))
                 .findAny()
                 .orElseGet(
-                        () -> Stream.of(this.testClassMembers.getStaticFields())
+                        () -> Stream.of(this.declarationDetails.getDeclaringTypeMembers().getStaticFields())
                                 .filter(field -> fieldName.equals(field.getName()))
                                 .findAny()
                                 .get());
-        return this.context.getTypeContext().createFieldScope(resolvedField, this.testClassMembers);
+        return this.context.getTypeContext().createFieldScope(resolvedField, this.declarationDetails);
     }
 
     protected MethodScope getTestClassMethod(String methodName) {
-        ResolvedMethod resolvedMethod = Stream.of(this.testClassMembers.getMemberMethods())
+        ResolvedMethod resolvedMethod = Stream.of(this.declarationDetails.getDeclaringTypeMembers().getMemberMethods())
                 .filter(method -> methodName.equals(method.getName()))
                 .findAny()
                 .orElseGet(
-                        () -> Stream.of(this.testClassMembers.getStaticMethods())
+                        () -> Stream.of(this.declarationDetails.getDeclaringTypeMembers().getStaticMethods())
                                 .filter(method -> methodName.equals(method.getName()))
                                 .findAny()
                                 .get());
-        return this.context.getTypeContext().createMethodScope(resolvedMethod, this.testClassMembers);
+        return this.context.getTypeContext().createMethodScope(resolvedMethod, this.declarationDetails);
     }
 }