+++ jsonschema-module-swagger-1.5/src/test/java/com/github/victools/jsonschema/module/swagger15/TestType.java
@@ -21,6 +21,7 @@ import com.fasterxml.classmate.ResolvedTypeWithMembers;
 import com.fasterxml.classmate.members.ResolvedField;
 import com.fasterxml.classmate.members.ResolvedMethod;
 import com.github.victools.jsonschema.generator.FieldScope;
+import com.github.victools.jsonschema.generator.MemberScope;
 import com.github.victools.jsonschema.generator.MethodScope;
 import com.github.victools.jsonschema.generator.TypeContext;
 import com.github.victools.jsonschema.generator.impl.TypeContextFactory;
@@ -32,44 +33,43 @@ import java.util.stream.Stream;
 public class TestType {
 
     private final TypeContext context;
-    private final ResolvedType resolvedTestClass;
-    private final ResolvedTypeWithMembers testClassMembers;
+    private final MemberScope.DeclarationDetails declarationDetails;
 
     public TestType(Class<?> testClass) {
         this.context = TypeContextFactory.createDefaultTypeContext();
-        this.resolvedTestClass = this.context.resolve(testClass);
-        this.testClassMembers = this.context.resolveWithMembers(this.resolvedTestClass);
+        ResolvedType resolvedTestClass = this.context.resolve(testClass);
+        this.declarationDetails = new MemberScope.DeclarationDetails(resolvedTestClass, this.context.resolveWithMembers(resolvedTestClass));
     }
 
     public FieldScope getMemberField(String fieldName) {
-        return this.getField(this.testClassMembers.getMemberFields(), fieldName);
+        return this.getField(this.declarationDetails.getDeclaringTypeMembers().getMemberFields(), fieldName);
     }
 
     public FieldScope getStaticField(String fieldName) {
-        return this.getField(this.testClassMembers.getStaticFields(), fieldName);
+        return this.getField(this.declarationDetails.getDeclaringTypeMembers().getStaticFields(), fieldName);
     }
 
     private FieldScope getField(ResolvedField[] fields, String fieldName) {
         return Stream.of(fields)
                 .filter(field -> fieldName.equals(field.getName()))
                 .findAny()
-                .map(field -> this.context.createFieldScope(field, this.testClassMembers))
+                .map(field -> this.context.createFieldScope(field, this.declarationDetails))
                 .get();
     }
 
     public MethodScope getMemberMethod(String methodName) {
-        return this.getMethod(this.testClassMembers.getMemberMethods(), methodName);
+        return this.getMethod(this.declarationDetails.getDeclaringTypeMembers().getMemberMethods(), methodName);
     }
 
     public MethodScope getStaticMethod(String methodName) {
-        return this.getMethod(this.testClassMembers.getStaticMethods(), methodName);
+        return this.getMethod(this.declarationDetails.getDeclaringTypeMembers().getStaticMethods(), methodName);
     }
 
     private MethodScope getMethod(ResolvedMethod[] methods, String methodName) {
         return Stream.of(methods)
                 .filter(method -> methodName.equals(method.getName()))
                 .findAny()
-                .map(method -> this.context.createMethodScope(method, this.testClassMembers))
+                .map(method -> this.context.createMethodScope(method, this.declarationDetails))
                 .get();
     }
 }